Fix the errors in more external layers after domain refactoring. Fix interactors and set up `CommandEnvelope`s. Remove some legacy CRUD repositories.
Changes have touched Application, Infra and Web layers.

+++ App.Domain/Shared/Random.fs
@@ -4,5 +4,5 @@ type IRandom =
     abstract member ShuffleList<'a> : seed: int -> list: 'a list -> 'a list
     /// From Min (inclusive) to Max (inclusive)
     abstract member NextInt: Min: int * Max: int -> int
-    abstract member NextDouble: unit -> uint64
+    abstract member NextDouble: unit -> double
     abstract member NextUInt64: unit -> uint64
